% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_hat_V.R
\name{gen_hat_sigma_squared_panel_fast}
\alias{gen_hat_sigma_squared_panel_fast}
\title{Generate hat_sigma_squared for Panel Data}
\usage{
gen_hat_sigma_squared_panel_fast(precomp)
}
\arguments{
\item{precomp}{List from precompute_panel_all}
}
\value{
hat_sigma_squared (T_total x 1 matrix)
}
\description{
Formula: hat_sigma^2_t = (1/n) sum_i pi_i^2 * psi_{i,t,theta}^2
where psi_{i,t,theta} = beta_i * (psi*\emph{{i,t} - sum_s lambda_s psi*}{i,s})
}
