% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_hat_V.R
\name{gen_hat_sigma_squared_RC}
\alias{gen_hat_sigma_squared_RC}
\title{Generate hat_sigma_squared for Repeated Cross-Sections}
\usage{
gen_hat_sigma_squared_RC(w, data, lambda, group_means, n_jt, K, n, T_star, L)
}
\description{
For RC, individual i appears at exactly ONE time period t_i.
The differenced influence function at period t for individual i at t_i is:
psi_{i,t} = (1 - lambda_{t_i}) * psi*\emph{i   if t = t_i
psi}{i,t} = -lambda_{t_i} * psi*_i        if t != t_i (and t_i is pre-treatment)
}
\details{
Therefore sigma_t^2 includes contributions from ALL pre-treatment individuals:
sigma_t^2 = (1/n) * [(1-lambda_t)^2 * sum_vec\link{t} + sum_{s!=t} lambda_s^2 * sum_vec\link{s}]

This expands to:
sigma_t^2 = (1/n) * [total_const + (1 - 2*lambda_t) * sum_vec\link{t}]
where total_const = sum_s lambda_s^2 * sum_vec\link{s}
}
