% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_hat_V.R
\name{gen_hat_V_panel_fast}
\alias{gen_hat_V_panel_fast}
\title{Generate hat_V for Panel Data}
\usage{
gen_hat_V_panel_fast(precomp)
}
\arguments{
\item{precomp}{List from precompute_panel_all}
}
\value{
hat_V matrix ((K-1) x (K-1))
}
\description{
Generate hat_V for Panel Data
}
