% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_hat_V.R
\name{gen_hat_V_RC}
\alias{gen_hat_V_RC}
\title{Generate hat_V for Repeated Cross-Sections}
\usage{
gen_hat_V_RC(
  w,
  hat_Gamma_pre,
  data,
  lambda,
  B2,
  group_means,
  n_jt,
  K,
  n,
  T_star
)
}
\description{
Formula (Line 947-948):
hat_V_RC(w) = (1/T_0) sum_t { (1/n) sum_i (sum_j w_j psi*\emph{{ij,t} - psi*}{i0,t})^2
* B2' M_t B2 }
where M_t = (1/T_0) mu_t mu_t' - lambda_t (bar_mu mu_t' + mu_t bar_mu')
+ T_0 lambda_t^2 bar_mu bar_mu'
}
