% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_C.R
\name{in_C_osqp}
\alias{in_C_osqp}
\title{Check Membership in Confidence Set C_{1-kappa}}
\usage{
in_C_osqp(w, kappa, hat_H, hat_h, precomp, n, K, tolerance)
}
\arguments{
\item{w}{Vector of weights (K x 1).}

\item{kappa}{Kappa level.}

\item{hat_H}{hat_H matrix.}

\item{hat_h}{hat_h vector.}

\item{precomp}{Precomputed matrix B2 \%\emph{\% solve(hat_V) \%}\% t(B2).}

\item{n}{Total sample size.}

\item{K}{Number of untreated groups.}

\item{tolerance}{Numerical tolerance.}
}
\value{
Logical: TRUE if in C, FALSE otherwise.
}
\description{
The optimization problem:
min_{r} (phi - r)' P (phi - r)  subject to w'r = 0, r >= 0
}
\details{
Rewritten in standard OSQP form:
min_{r} (1/2) r' (2P) r - (2P*phi)' r
subject to: \link{w'; -I} r <= \link{0; 0}  and  \link{w'; I} r >= \link{0; 0}
i.e., w'r = 0, r >= 0
}
