% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_hat_V.R
\name{precompute_panel_all}
\alias{precompute_panel_all}
\title{Precompute all panel quantities}
\usage{
precompute_panel_all(
  w,
  data,
  hat_Gamma_pre,
  lambda,
  B2,
  group_means,
  n_jt,
  K,
  n,
  T_star,
  T_total
)
}
\arguments{
\item{w}{Vector of weights (K x 1).}

\item{data}{Full data frame sorted by (id, t).}

\item{hat_Gamma_pre}{Pre-treatment DIFFERENCED means matrix (T_0 x K).}

\item{lambda}{Vector of lambdas (T_0).}

\item{B2}{Orthonormal basis matrix (K x (K-1)).}

\item{group_means}{Matrix of group means ((K+1) x T_total).}

\item{n_jt}{Matrix of weighted sums ((K+1) x T_total).}

\item{K}{Number of donors.}

\item{n}{Total sample size.}

\item{T_star}{First post-treatment period.}

\item{T_total}{Total periods.}
}
\value{
List with all precomputed quantities
}
\description{
Precompute all panel quantities
}
